+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project.
 ### Added
 - `orjson` support version `3.6.8` and above.
 - Adapter classes to be able to explicitly select a json parser.
+- The fields `join_to_send_messages` and `join_by_request` to the class `Chat`.
 
 ### Changed
 - The minimum required version of `aiohttp` is now `3.8.1`.